re-organize structure
