+++ src/App.js
@@ -2,14 +2,13 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "./pages/Home";
 import Cart from "./pages/Cart";
 import Header from "./components/Header";
-import Language from "./components/Language";
-import SignIn from "./components/SignIn";
-import SignUp from "./components/SignUp";
-import Deal from "./components/Deal";
-import CustomerService from "./components/CustomerService";
-import Registry from "./components/Registry";
-import GiftCard from "./components/GiftCard";
-import Sell from "./components/Sell";
+import SignIn from "./pages/SignIn";
+import SignUp from "./pages/SignIn/SignUp";
+import Deal from "./pages/Deal";
+import CustomerService from "./pages/CustomerService";
+import Registry from "./pages/Registry";
+import GiftCard from "./pages/GiftCard";
+import Sell from "./pages/Sell";
 
 function App() {
   return (
@@ -18,7 +17,6 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/cart" element={<Cart />} />
-        <Route path="/language" element={<Language />} />
         <Route path="/signin" element={<SignIn />} />
         <Route path="/signup" element={<SignUp />} />
         <Route path="/deal" element={<Deal />} />
