--- src/components/Language.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Language = () => {
-  return (
-    <div>Language</div>
-  )
-}
-
-export default Language
